requirements.txt
python-dotenv
openai
gradio
gTTS
